requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2
python-dotenv
pymysql
boto3
python-multipart
moviepy
alembic
uuid